Try to fix aot crash when doing performance test
When inst/uninst repeatedly, the app server will crash.

We have already fixed it for wasm (interpreter mode), but AOT mode is
not fixed by this patch.

There seems to be other issue when visiting memory, i.e., AOT uses mmap
to allocate TEXT (code) memory, because the memory need to be
executable. But in another thread, there might be an async call to
munmap, causing the memory to be freed.

This is only a doubt, I havn't proven it yet.